pyproject.toml
[project]
name="hyperlocal-platform"
version = "2.0.1"
dependencies=[
    "pyjwt",
    "cryptography",
    "argon2-cffi",
    "orjson",
    "icecream",
    "pytz",
    "pydantic",
    "redis",
    "sqlalchemy"
]